screenreader/requirements.txt
pyautogui
gTTS
google-generativeai
Pillow
keyboard